[#1546] Separate class methods
